+++ cuda/kernels.cu
@@ -21,21 +21,21 @@ __constant__ uint64_t pow4_T[33]={3*1L, 3*4L, 3*16L, 3*64L, 3*256L, 3*1024L, 3*4
 
 #define MIN_P_IDENT 80
 
-__global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, uint64_t * left_offset, uint64_t * right_offset, const char * seq_x, const char * seq_y, uint64_t query_len, uint64_t ref_len, uint64_t x_seq_off, uint64_t y_seq_off, uint64_t x_lim, uint64_t y_lim){
+__global__ void kernel_frags_forward_register(uint32_t * h_p1, uint32_t * h_p2, uint32_t * left_offset, uint32_t * right_offset, const char * seq_x, const char * seq_y, uint32_t query_len, uint32_t ref_len, uint32_t x_seq_off, uint32_t y_seq_off, uint32_t x_lim, uint32_t y_lim){
 
 
 	// LEFT ALIGNMENT
-	int64_t hp1 = (int64_t) h_p1[blockIdx.x];
-	int64_t warp_pos_x_left = hp1;
-	int64_t warp_pos_y_left = (int64_t) h_p2[blockIdx.x];
-	int64_t thre_pos_x, thre_pos_y;
-	uint64_t best_offset_left = (uint64_t) warp_pos_x_left;
-	int64_t score = 32, best_score = 32;
+	int32_t hp1 = (int32_t) h_p1[blockIdx.x];
+	int32_t warp_pos_x_left = hp1;
+	int32_t warp_pos_y_left = (int32_t) h_p2[blockIdx.x];
+	int32_t thre_pos_x, thre_pos_y;
+	uint32_t best_offset_left = (uint32_t) warp_pos_x_left;
+	int32_t score = 32, best_score = 32;
 	int p_ident = 100;
 	int cell_score;
 	
 	//while(p_ident > MIN_P_IDENT && (warp_pos_x_left - 32) >= (int64_t) x_seq_off && (warp_pos_y_left - 32) >= (int64_t) y_seq_off)
-	while(score > 0 && (warp_pos_x_left - 32) >= (int64_t) x_seq_off && (warp_pos_y_left - 32) >= (int64_t) y_seq_off)
+	while(score > 0 && (warp_pos_x_left - 32) >= (int32_t) x_seq_off && (warp_pos_y_left - 32) >= (int32_t) y_seq_off)
 	{
 		
 		warp_pos_x_left -= 32;
@@ -45,8 +45,8 @@ __global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, 
 
 		
 
-		char v_x = seq_x[thre_pos_x - (int64_t) x_seq_off];
-		char v_y = seq_y[thre_pos_y - (int64_t) y_seq_off];
+		char v_x = seq_x[thre_pos_x - (int32_t) x_seq_off];
+		char v_y = seq_y[thre_pos_y - (int32_t) y_seq_off];
 
 		//cell_score = (v_x == v_y && v_x != '\0' && v_y != '\0');
 		cell_score = (v_x == v_y);
@@ -56,8 +56,8 @@ __global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, 
 
 		
 		int idents = __shfl_sync(0xFFFFFFFF, cell_score, 0);
-		score = score + (int64_t) idents;
-		score = score - (int64_t) (32 - idents);
+		score = score + (int32_t) idents;
+		score = score - (int32_t) (32 - idents);
 		p_ident = ((100 * score) / (int) (hp1 - warp_pos_x_left));
 		if(score > best_score && p_ident >= MIN_P_IDENT){ best_score = score; best_offset_left = warp_pos_x_left; }
 		
@@ -69,21 +69,21 @@ __global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, 
 	
 
 	// RIGHT ALIGNMENT
-	int64_t warp_pos_x_right = hp1 + 32;
-	int64_t warp_pos_y_right = (int64_t) h_p2[blockIdx.x] + 32;
-	uint64_t best_offset_right = (uint64_t) warp_pos_x_right;
+	int32_t warp_pos_x_right = hp1 + 32;
+	int32_t warp_pos_y_right = (int32_t) h_p2[blockIdx.x] + 32;
+	uint32_t best_offset_right = (uint32_t) warp_pos_x_right;
 	score = 32;
 	best_score = 32;
 	p_ident = 100;
 
 	
 	//while(p_ident > MIN_P_IDENT && (warp_pos_x_right + 32) < (int64_t) x_lim && (warp_pos_y_right + 32) < (int64_t) y_lim)
-	while(score > 0 && (warp_pos_x_right + 32) < (int64_t) x_lim && (warp_pos_y_right + 32) < (int64_t) y_lim)
+	while(score > 0 && (warp_pos_x_right + 32) < (int32_t) x_lim && (warp_pos_y_right + 32) < (int32_t) y_lim)
 	{
 		thre_pos_x = warp_pos_x_right + threadIdx.x;
 		thre_pos_y = warp_pos_y_right + threadIdx.x;
-		char v_x = seq_x[thre_pos_x - (int64_t) x_seq_off];
-		char v_y = seq_y[thre_pos_y - (int64_t) y_seq_off];
+		char v_x = seq_x[thre_pos_x - (int32_t) x_seq_off];
+		char v_y = seq_y[thre_pos_y - (int32_t) y_seq_off];
 
 		//cell_score = (v_x == v_y && v_x != '\0' && v_y != '\0');
 		cell_score = (v_x == v_y);
@@ -92,9 +92,9 @@ __global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, 
 			cell_score += __shfl_down_sync(0xFFFFFFFF, cell_score, offset);
 		
 		int idents = __shfl_sync(0xFFFFFFFF, cell_score, 0);
-		score = score + (int64_t) idents;
-		score = score - (int64_t) (32 - idents);
-		p_ident = (int) ((100 * score) / (int64_t) (warp_pos_x_right - (hp1)));
+		score = score + (int32_t) idents;
+		score = score - (int32_t) (32 - idents);
+		p_ident = (int) ((100 * score) / (int32_t) (warp_pos_x_right - (hp1)));
 
 		warp_pos_x_right += 32;
 		warp_pos_y_right += 32; 
@@ -107,30 +107,29 @@ __global__ void kernel_frags_forward_register(uint64_t * h_p1, uint64_t * h_p2, 
 
 	// Save at the end
 	if(threadIdx.x == 0){
-		left_offset[blockIdx.x] = hp1 - (uint64_t) best_offset_left;
-		//right_offset[blockIdx.x] = (uint64_t) (best_offset_right + 32) - h_p1[blockIdx.x];
-		right_offset[blockIdx.x] = (uint64_t) (best_offset_right) - hp1;
+		left_offset[blockIdx.x] = hp1 - (uint32_t) best_offset_left;
+		right_offset[blockIdx.x] = (uint32_t) (best_offset_right) - hp1;
 	}
 
 }
 
 
-__global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, uint64_t * left_offset, uint64_t * right_offset, const char * seq_x, const char * seq_y, uint64_t query_len, uint64_t ref_len, uint64_t x_seq_off, uint64_t y_seq_off, uint64_t x_lim, uint64_t y_lim){
+__global__ void kernel_frags_reverse_register(uint32_t * h_p1, uint32_t * h_p2, uint32_t * left_offset, uint32_t * right_offset, const char * seq_x, const char * seq_y, uint32_t query_len, uint32_t ref_len, uint32_t x_seq_off, uint32_t y_seq_off, uint32_t x_lim, uint32_t y_lim){
 
 	// LEFT ALIGNMENT
-	int64_t hp1 = (int64_t) h_p1[blockIdx.x];
-	int64_t warp_pos_x_left = hp1;
-	int64_t warp_pos_y_left = (int64_t) h_p2[blockIdx.x];
-	int64_t thre_pos_x, thre_pos_y;
-	uint64_t best_offset_left = (uint64_t) warp_pos_x_left;
-	int64_t score = 32, best_score = 32;
+	int32_t hp1 = (int32_t) h_p1[blockIdx.x];
+	int32_t warp_pos_x_left = hp1;
+	int32_t warp_pos_y_left = (int32_t) h_p2[blockIdx.x];
+	int32_t thre_pos_x, thre_pos_y;
+	uint32_t best_offset_left = (uint32_t) warp_pos_x_left;
+	int32_t score = 32, best_score = 32;
 	int p_ident = 100;
 	int cell_score;
 	
 
 	
 	//while(p_ident > MIN_P_IDENT && (warp_pos_x_left - 32) >= (int64_t) x_seq_off && (warp_pos_y_left - 32) >= (int64_t) y_seq_off)
-	while(score > 0 && (warp_pos_x_left - 32) >= (int64_t) x_seq_off && (warp_pos_y_left - 32) >= (int64_t) y_seq_off)
+	while(score > 0 && (warp_pos_x_left - 32) >= (int32_t) x_seq_off && (warp_pos_y_left - 32) >= (int32_t) y_seq_off)
 	{
 		
 		warp_pos_x_left -= 32;
@@ -139,8 +138,8 @@ __global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, 
 		thre_pos_y = warp_pos_y_left + threadIdx.x;
 
 		
-		char v_x = seq_x[thre_pos_x - (int64_t) x_seq_off];
-		char v_y = seq_y[thre_pos_y - (int64_t) y_seq_off];
+		char v_x = seq_x[thre_pos_x - (int32_t) x_seq_off];
+		char v_y = seq_y[thre_pos_y - (int32_t) y_seq_off];
 
 		
 		//if(blockIdx.x == 0) printf("T%d -> %c - %c\n", threadIdx.x, v_x, v_y);
@@ -154,8 +153,8 @@ __global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, 
 
 		
 		int idents = __shfl_sync(0xFFFFFFFF, cell_score, 0);
-		score = score + (int64_t) idents;
-		score = score - (int64_t) (32 - idents);
+		score = score + (int32_t) idents;
+		score = score - (int32_t) (32 - idents);
 		p_ident = ((100 * score) / (int) (hp1 - warp_pos_x_left));
 		if(score > best_score && p_ident >= MIN_P_IDENT){ best_score = score; best_offset_left = warp_pos_x_left; }
 		
@@ -169,21 +168,21 @@ __global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, 
 	
 
 	// RIGHT ALIGNMENT
-	int64_t warp_pos_x_right = hp1 + 32;
-	int64_t warp_pos_y_right = (int64_t) h_p2[blockIdx.x] + 32;
-	uint64_t best_offset_right = (uint64_t) warp_pos_x_right;
+	int32_t warp_pos_x_right = hp1 + 32;
+	int32_t warp_pos_y_right = (int32_t) h_p2[blockIdx.x] + 32;
+	uint32_t best_offset_right = (uint32_t) warp_pos_x_right;
 	score = 32;
 	best_score = 32;
 	p_ident = 100;
 
 	
 	//while(p_ident > MIN_P_IDENT && (warp_pos_x_right + 32) < (int64_t) x_lim && (warp_pos_y_right + 32) < (int64_t) y_lim)
-	while(score > 0 && (warp_pos_x_right + 32) < (int64_t) x_lim && (warp_pos_y_right + 32) < (int64_t) y_lim)
+	while(score > 0 && (warp_pos_x_right + 32) < (int32_t) x_lim && (warp_pos_y_right + 32) < (int32_t) y_lim)
 	{
 		thre_pos_x = warp_pos_x_right + threadIdx.x;
 		thre_pos_y = warp_pos_y_right + threadIdx.x;
-		char v_x = seq_x[thre_pos_x - (int64_t) x_seq_off];
-		char v_y = seq_y[thre_pos_y - (int64_t) y_seq_off];
+		char v_x = seq_x[thre_pos_x - (int32_t) x_seq_off];
+		char v_y = seq_y[thre_pos_y - (int32_t) y_seq_off];
 
 		//cell_score = (v_x == v_y && v_x != '\0' && v_y != '\0');
 		cell_score = (v_x == v_y);
@@ -192,9 +191,9 @@ __global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, 
 			cell_score += __shfl_down_sync(0xFFFFFFFF, cell_score, offset);
 		
 		int idents = __shfl_sync(0xFFFFFFFF, cell_score, 0);
-		score = score + (int64_t) idents;
-		score = score - (int64_t) (32 - idents);
-		p_ident = (int) ((100 * score) / (int64_t) (warp_pos_x_right - (hp1)));
+		score = score + (int32_t) idents;
+		score = score - (int32_t) (32 - idents);
+		p_ident = (int) ((100 * score) / (int32_t) (warp_pos_x_right - (hp1)));
 
 		warp_pos_x_right += 32;
 		warp_pos_y_right += 32; 
@@ -208,9 +207,9 @@ __global__ void kernel_frags_reverse_register(uint64_t * h_p1, uint64_t * h_p2, 
 
 	// Save at the end
 	if(threadIdx.x == 0){
-		left_offset[blockIdx.x] = hp1 - (uint64_t) best_offset_left;
+		left_offset[blockIdx.x] = hp1 - (uint32_t) best_offset_left;
 		//right_offset[blockIdx.x] = (uint64_t) (best_offset_right + 32) - h_p1[blockIdx.x];
-		right_offset[blockIdx.x] = (uint64_t) (best_offset_right) - hp1;
+		right_offset[blockIdx.x] = (uint32_t) (best_offset_right) - hp1;
 	}
 
 }
@@ -396,7 +395,7 @@ __global__ void kernel_register_fast_hash_rotational(uint64_t * hashes, uint64_t
 	
 }
 
-__global__ void kernel_index_global32(uint64_t * hashes, uint64_t * positions, const char * sequence, uint64_t offset) {
+__global__ void kernel_index_global32(uint64_t * hashes, uint32_t * positions, const char * sequence, uint32_t offset) {
 		
 
 	uint64_t k, hash = 0;
@@ -423,7 +422,7 @@ __global__ void kernel_index_global32(uint64_t * hashes, uint64_t * positions, c
 	positions[threadIdx.x + blockIdx.x * blockDim.x] = (threadIdx.x + blockIdx.x * blockDim.x + offset) | (~bad);
 }
 
-__global__ void kernel_reverse_complement(const char * sequence, char * reverse_sequence, uint64_t seq_len) {
+__global__ void kernel_reverse_complement(const char * sequence, char * reverse_sequence, uint32_t seq_len) {
 	
 	//uint64_t id = (31 - threadIdx.x) + blockIdx.x * blockDim.x;
 	uint64_t id = threadIdx.x + blockIdx.x * blockDim.x;
@@ -432,7 +431,7 @@ __global__ void kernel_reverse_complement(const char * sequence, char * reverse_
 	if(id < seq_len){
 
 		//uint64_t lookup = seq_len - id;
-		uint64_t lookup = (seq_len - 1) - id;
+		uint32_t lookup = (seq_len - 1) - id;
 		char original = sequence[id];
 		char complement = 'N';
 		if(original == 'A') complement = 'T';
